lint: add golangci-lint
This change enables golangci-lint with suite of linters enabled.
We will add additional linters one-by-one.

The linter is currently only enabled in CI because it takes
quite some time to run.

GitOrigin-RevId: e85290c24e6f4e9439196496a258d4ea70aaeb88

+++ go/lib/snet/packet.go
@@ -344,8 +344,8 @@ func (p *Packet) Decode() error {
 	switch l4 {
 	case slayers.LayerTypeSCIONUDP:
 		p.Payload = UDPPayload{
-			SrcPort: uint16(udpLayer.SrcPort),
-			DstPort: uint16(udpLayer.DstPort),
+			SrcPort: udpLayer.SrcPort,
+			DstPort: udpLayer.DstPort,
 			Payload: udpLayer.Payload,
 		}
 	case slayers.LayerTypeSCMP:
